Purchases: Extract CreditCardPage component
