+++ assets/stylesheets/_components.scss
@@ -179,12 +179,12 @@
 @import 'me/profile-links/style';
 @import 'me/purchases/cancel-private-registration/style';
 @import 'me/purchases/cancel-purchase/style';
+@import 'me/purchases/components/credit-card-form/style';
 @import 'me/purchases/components/loading-placeholder/style';
 @import 'me/purchases/confirm-cancel-domain/style';
 @import 'me/purchases/list/item/style';
 @import 'me/purchases/list/site/style';
 @import 'me/purchases/manage-purchase/style';
-@import 'me/purchases/payment/edit-card-details/style';
 @import 'me/purchases/remove-purchase/style';
 @import 'me/reauth-required/style';
 @import 'me/security-2fa-backup-codes-list/style';
